Refactor issue reporters

+++ internal/issue/issue_service.go
@@ -28,7 +28,7 @@ type IssueStore interface {
 }
 
 type Issue struct {
-	PageReportId int
+	PageReportId int64
 	CrawlId      int64
 	ErrorType    int
 }
